backend/requirements.txt
Flask==2.0.3
Flask-CORS==3.0.10
Werkzeug==2.0.3
pymysql==1.0.2
neo4j==5.19.0
pyjwt==2.4.0
bcrypt==4.0.1
cryptography==42.0.8
python-dotenv==0.20.0
requests==2.28.2
beautifulsoup4==4.11.1
gunicorn==23.0.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
xgboost==2.1.4
openpyxl==3.1.5
